added comments + made some things private + moved others around
